% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_interview_md.R
\name{upload_interview_md}
\alias{upload_interview_md}
\title{Make markdown files for interviews}
\usage{
upload_interview_md(id = "1HPQDH3tOXtZb1DV--8wR9CKAzUz5aywWc2vM3OQ5SrU")
}
\arguments{
\item{id}{Your Google Sheet id}
}
\description{
This function will create markdown files in the Hugo Apero style for interviews
specified in a Google Sheet.
}
