% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_paper_md.R
\name{upload_paper_md}
\alias{upload_paper_md}
\title{Make markdown files for papers}
\usage{
upload_paper_md(id = "1HPQDH3tOXtZb1DV--8wR9CKAzUz5aywWc2vM3OQ5SrU")
}
\arguments{
\item{id}{Your Google Sheet id}
}
\description{
This function will create markdown files in the Hugo Apero style for papers
specified in a Google Sheet.
}
