% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_talk_md.R
\name{upload_talk_md}
\alias{upload_talk_md}
\title{Make markdown files for talks}
\usage{
upload_talk_md(id = "1-PItelqpv0Sb_LdiEDqb8O3D_Roii5nVTL07IRVbRtA")
}
\arguments{
\item{id}{Your Google Sheet id}
}
\description{
This function will create markdown files in the Hugo Apero style for talks
specified in a Google Sheet.
}
