% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_workshop_md.R
\name{upload_workshop_md}
\alias{upload_workshop_md}
\title{Make markdown files for workshops}
\usage{
upload_workshop_md(
  id = "1-PItelqpv0Sb_LdiEDqb8O3D_Roii5nVTL07IRVbRtA",
  filter_column = "what",
  filter_term = "Workshop"
)
}
\arguments{
\item{id}{Your Google Sheet id}

\item{filter_column}{column to filter sheet by}

\item{filter_term}{term to filter column specified in `filter_column`}
}
\description{
This function will create markdown files in the Hugo Apero style for talks
specified in a Google Sheet after filtering (in this case to workshops).
}
